+++ src/utils/json-utils/json-util.js
@@ -1,11 +1,10 @@
 // @flow
 import camelCase from "lodash/camelCase";
-import curry from "lodash/curry";
 import forEach from "lodash/forEach";
 import isObject from "lodash/isObject";
 import upperFirst from "lodash/upperFirst";
 
-export const convertObjectKeys = (keyMutator: Function, ignoredKeys: Array<string>, object: any) => {
+export const convertObjectKeys = (keyMutator: Function, ignoredKeys?: Array<string>, object: any): any => {
   let convertedObject;
   if (Array.isArray(object)) {
     convertedObject = [];
@@ -18,7 +17,7 @@ export const convertObjectKeys = (keyMutator: Function, ignoredKeys: Array<strin
   } else if (isObject(object)) {
     convertedObject = {};
     forEach(object, (value, key) => {
-      if (typeof value === "object" && !ignoredKeys.includes(key)) {
+      if (typeof value === "object" && (!ignoredKeys || !ignoredKeys.includes(key))) {
         value = convertObjectKeys(keyMutator, ignoredKeys, value);
       }
       if (/^[A-Z0-9_]*$/.test(key) || (/^[a-z0-9_]*$/.test(key) && key.includes("_"))) {
@@ -35,20 +34,11 @@ export const convertObjectKeys = (keyMutator: Function, ignoredKeys: Array<strin
   return convertedObject;
 };
 
-const ignoredKeys = [
-  "InputState",
-  "inputState",
-  "OutputState",
-  "outputState",
-  "ProgramStateUpdate",
-  "programStateUpdate",
-];
-// $FlowFixMe This type is incompatible with function type Callable
-export const camelizeObjectKeys: Function = curry(convertObjectKeys)(camelCase, ignoredKeys);
-
 const titleize = key => upperFirst(camelCase(key));
-// $FlowFixMe This type is incompatible with function type Callable
-export const titleizeObjectKeys: Function = curry(convertObjectKeys)(titleize, ignoredKeys);
+// eslint-disable-next-line max-len
+export const titleizeObjectKeys = (object: any, ignoredKeys?: Array<string>) => convertObjectKeys(titleize, ignoredKeys, object);
+// eslint-disable-next-line max-len
+export const camelizeObjectKeys = (object: any, ignoredKeys?: Array<string>) => convertObjectKeys(camelCase, ignoredKeys, object);
 
 // warning: does not work if the map's key or value is another map object.
 export const mapToJson: Function = (map: Map<*, *>) => {