+++ src/index.js
@@ -136,4 +136,3 @@ export {
 } from "./utils/url-utils";
 export { Warning } from "./utils/warning-utils";
 export { toDelimitedReport, toTableRow } from "./utils/table-utils";
-export { processCommandLineArgs } from "./utils/command-line";