+++ src/index.js
@@ -17,6 +17,7 @@ export {
   grailPurpleDark,
   grailPurpleLight,
   grailPurpleMed,
+  grailPurpleActionColor,
   grailRed,
   grailTan,
   white,
@@ -77,9 +78,16 @@ export {
   trimObjectValues,
 } from "./utils/json-utils";
 export {
-  buildDateRangeString, extractDateRange, formatDate, formatDateTime,
+  buildDateRangeString,
+  extractDateRange,
+  formatDate,
+  formatDateTime,
 } from "./utils/date-utils";
-export { getColHeader, getGridCellPosition, getRowHeader } from "./utils/grid-utils";
+export {
+  getColHeader,
+  getGridCellPosition,
+  getRowHeader,
+} from "./utils/grid-utils";
 export {
   getItemsFromOmniValue,
   getKeysForSearchDef,
@@ -91,7 +99,10 @@ export {
   parseValuesFromOmniText,
 } from "./utils/omni-search-util";
 export {
-  generateFilledArray, serializePromises, mapBy, toPairWise,
+  generateFilledArray,
+  serializePromises,
+  mapBy,
+  toPairWise,
 } from "./utils/array-utils";
 export { hasKey, hasKeyValue, setKeyValue } from "./utils/key-value-utils";
 export { fnv1 } from "./utils/hash-utils";
@@ -118,7 +129,10 @@ export {
   valuesToSuggestions,
 } from "./utils/string-utils";
 export {
-  sanitizeId, getInputBarcode, getSampleLabel, getSamplePrefix,
+  sanitizeId,
+  getInputBarcode,
+  getSampleLabel,
+  getSamplePrefix,
 } from "./utils/id-utils";
 export {
   expandSearchValues,