+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 3. Add grid util functions: `getColHeader`, `getRowHeader`, and `getGridCellPosition`.
 4. Add type `mapValuesDispatcher` function to allow coders the option of dispatching
    value mapper in Omni Search fields.
+5. Add `parseCommandLineArguments` utility to handle common node command
+   line argument parsing.
 
 ## 0.10.0 - 2019-04-01