+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 4. Add type `mapValuesDispatcher` function to allow coders the option of dispatching
    value mapper in Omni Search fields.
 5. Move node server lib code like `passport` utils and `commandLine` utils to `@grail/server-lib`.
+6. Add `grailPurpleActionColor`. (Intended for links and actions).
 
 ## 0.10.0 - 2019-04-01